+++ frontend/rush/src/components/home/button/groupMap/GroupContent.js
