+++ frontend/rush/src/components/home/button/groupMap/GroupList.js
